Make the PipeSpawner collision box much smaller to prevent unintended pipe spawns
The PipeSpawner's collision box was about half the height of the screen. One effect of this was sometimes both the top and bottom colliders off a pipe would collide with it, causign 2 pipes to spawn on top of eachother.

Fixed here by reducing the size, so its only possible for the top collider to collide with it.

A better alternative to look at in the future would be to give the entire Pipe a collision box, instead of using the top and bottom parts.